Refactor LicenseType template and modify build.rs
It now contains a try_from() to allow building from &str. Unfortunately
the TryFrom trait was not implemented because it is currently only
available on nightly.
LicenseTypeError was also added which is an enum that contains errors
related to LicenseType.
build.rs now also adds a new field called "abbreviation" to the parsed
JSON before sending it off to handlebars-rust. This is because the ls()
function should display the abbreviated name of the license as well as
the full name, and the JSON does not contain the abbreviated name as a
field, only as a key.